add Secret class
Subclass of PegSequence. Additional variable 'hidden' to make sure
the sequence is not displayed until the end of the game.
The to_s method also provides functionality for that.

When initialized, the sequence is filled with random Pegs chosen
from a given color range.

The secret class can evaluate a guess.